feat load model tools

+++ LearnOpenGL/utility/Mesh.cpp
@@ -32,7 +32,7 @@ void RMesh::Draw(Shader& shader)
 		{
 			number = std::to_string(specularNr++);
 		}
-		shader.setFloat(("material." + name + number).c_str(), i);
+		shader.setInteger(("material." + name + number).c_str(), i);
 		glBindTexture(GL_TEXTURE_2D, textures[i].id);
 	}
 	glActiveTexture(GL_TEXTURE0); // 通常默认激活的都是第0个纹理单元.
